Refactor class naming and add 'class' property
The class naming scheme was altered from 'ClassObject' to 'Object', and 'ClassPlayer' to 'Player' for simplicity and clarity. A property 'class' was introduced to differentiate between 'Object' and 'Player'. This property holds the class type name. The instances of these classes were also renamed to lower case. These changes contribute to more readable, structured and OOP-styled Lua code.